Move the createGeckoProfiler function into a helper file
